pyproject.toml
[project]
name = "semantic-kernel-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "semantic-kernel[all]",
    "semantic-kernel[azure]",
    "azure-identity"
]

